修复自提权更新问题

+++ tool/update-upgrade/ss-do-upgrade.sh
@@ -1,4 +1,8 @@
 #!/bin/bash
+if [ "$(id -u)" != "0" ] ; then
+	pkexec "$0" "$@"
+	exit
+fi
 trap "rm -f  /tmp/spark-store/upgradeStatus.txt" EXIT
 source /opt/durapps/spark-store/bin/bashimport/transhell.amber
 load_transhell_debug
